refactor: consolidate the two VLM batch-sort mixins into one module
RBLNVisionBatchSortMixin moves from qwen2_vl into
transformers/utils/generation_multimodal.py next to
RBLNMultimodalBatchSortMixin, and the duplicated forward guard becomes a
shared RBLNBatchSortGuardMixin._require_sorted_batch_inputs.
