+++ src/optimum/rbln/transformers/models/qwen2_vl/modeling_qwen2_vl.py
@@ -39,7 +39,7 @@ from ....modeling import RBLNModel
 from ....modeling_rope_utils import build_qwen_mrope_lookup, np_cos, np_sin, qwen_vit_rot_pos_ids
 from ....utils.logging import get_logger
 from ...modeling_outputs import _validate_output_hidden_states
-from ..decoderonly.generation_decoderonly import RBLNDecoderOnlyGenerationMixin, _permute_flat_segments
+from ...utils.generation_multimodal import RBLNVisionBatchSortMixin
 from ..decoderonly.modeling_decoderonly import (
     RBLNDecoderOnlyModel,
     RBLNDecoderOnlyModelForCausalLM,
@@ -61,113 +61,6 @@ if TYPE_CHECKING:
         AutoTokenizer,
         PretrainedConfig,
     )
-
-
-def _per_sample_patch_lens(grid_thw: torch.Tensor, rows_per_sample: list[int]) -> list[int]:
-    # patches for grid row i = t*h*w; sum rows owned by each sample
-    patches_per_row = grid_thw.prod(dim=-1).tolist()
-    lens, row_idx = [], 0
-    for n in rows_per_sample:
-        lens.append(sum(patches_per_row[row_idx : row_idx + n]))
-        row_idx += n
-    return lens
-
-
-class RBLNVisionBatchSortMixin:
-    # generate-entry batch sorting (use_batch_attn_opt) must carry the flattened vision
-    # inputs (patches / grid rows stacked on dim 0) along with their sample rows.
-    _video_grid_rows_are_chunks = False  # qwen3-style video grids are per temporal chunk
-    _generate_batch_sortable_kwargs = RBLNDecoderOnlyGenerationMixin._generate_batch_sortable_kwargs + (
-        "mm_token_type_ids",
-    )
-    _vision_sortable_keys = (
-        "pixel_values",
-        "pixel_values_videos",
-        "image_grid_thw",
-        "video_grid_thw",
-        "second_per_grid_ts",
-    )
-
-    def _sort_generation_inputs(
-        self, input_ids: torch.LongTensor | None, kwargs: dict
-    ) -> tuple[torch.LongTensor | None, torch.Tensor | None]:
-        presort_input_ids = input_ids
-        presort_mask = kwargs.get("attention_mask")
-        input_ids, unsort_idx = super()._sort_generation_inputs(input_ids, kwargs)
-        if unsort_idx is None or not any(kwargs.get(k) is not None for k in self._vision_sortable_keys):
-            return input_ids, unsort_idx
-        if presort_input_ids is None:
-            raise RuntimeError("Batch sorting flattened vision inputs requires input_ids.")
-
-        # per-sample counts come from the pre-sort rows; segments are then permuted to match
-        sort_idx = torch.argsort(unsort_idx)
-        image_rows, video_rows = self._vision_grid_rows_per_sample(presort_input_ids, presort_mask, kwargs)
-        self._sort_vision_kwargs(kwargs, sort_idx, image_rows, video_rows)
-        return input_ids, unsort_idx
-
-    def _vision_grid_rows_per_sample(
-        self, input_ids: torch.LongTensor, attention_mask: torch.Tensor | None, kwargs: dict
-    ) -> tuple[list[int], list[int]]:
-        # same counting as _preprocess_prefill: vision_start marker followed by image/video token
-        image_rows, video_rows = [], []
-        video_grid_thw = kwargs.get("video_grid_thw")
-        video_row_idx = 0
-        for b_idx in range(input_ids.shape[0]):
-            row = input_ids[b_idx]
-            if attention_mask is not None:
-                row = row[attention_mask[b_idx].bool()]
-            vision_start_indices = torch.argwhere(row == self.config.vision_start_token_id).squeeze(1)
-            vision_tokens = row[vision_start_indices + 1]
-            image_rows.append(int((vision_tokens == self.config.image_token_id).sum().item()))
-            video_nums = int((vision_tokens == self.config.video_token_id).sum().item())
-            if self._video_grid_rows_are_chunks and video_grid_thw is not None:
-                start_row = video_row_idx
-                consumed_video_chunks = 0
-                while video_row_idx < video_grid_thw.shape[0] and consumed_video_chunks < video_nums:
-                    consumed_video_chunks += int(video_grid_thw[video_row_idx, 0].item())
-                    video_row_idx += 1
-                video_rows.append(video_row_idx - start_row)
-            else:
-                video_rows.append(video_nums)
-        return image_rows, video_rows
-
-    def _sort_vision_kwargs(
-        self, kwargs: dict, sort_idx: torch.Tensor, image_rows: list[int], video_rows: list[int]
-    ) -> None:
-        for grid_key, pixel_key, seg_rows in (
-            ("image_grid_thw", "pixel_values", image_rows),
-            ("video_grid_thw", "pixel_values_videos", video_rows),
-        ):
-            grid = kwargs.get(grid_key)
-            if grid is None:
-                continue
-            pixels = kwargs.get(pixel_key)
-            if pixels is not None:
-                kwargs[pixel_key] = _permute_flat_segments(pixels, _per_sample_patch_lens(grid, seg_rows), sort_idx)
-            kwargs[grid_key] = _permute_flat_segments(grid, seg_rows, sort_idx)
-
-        second_per_grid_ts = kwargs.get("second_per_grid_ts")
-        if second_per_grid_ts is not None:
-            if isinstance(second_per_grid_ts, torch.Tensor):
-                kwargs["second_per_grid_ts"] = _permute_flat_segments(second_per_grid_ts, video_rows, sort_idx)
-            else:
-                bounds = [0]
-                for n in video_rows:
-                    bounds.append(bounds[-1] + n)
-                kwargs["second_per_grid_ts"] = [
-                    v for i in sort_idx.tolist() for v in second_per_grid_ts[bounds[i] : bounds[i + 1]]
-                ]
-
-    def _check_batch_inputs_sorted(self, batch_input: torch.Tensor | None, inputs_sorted: bool) -> None:
-        # this forward drives prefill_decoder per batch_idx itself, so the base forward's
-        # sorting fallback never runs; an unsorted direct call would silently mis-lay the KV cache
-        if inputs_sorted or batch_input is None or batch_input.shape[0] <= 1 or not self._batch_sort_enabled:
-            return
-        raise RuntimeError(
-            "use_batch_attn_opt requires batch inputs sorted by sequence length (descending). "
-            "Call generate(), which sorts and unsorts automatically, or pass `inputs_sorted=True` "
-            "with inputs (including flattened vision inputs) already sorted."
-        )
 
 
 class RBLNQwen2VisionTransformerPretrainedModel(RBLNModel):
@@ -746,7 +639,7 @@ class RBLNQwen2VLForConditionalGeneration(RBLNVisionBatchSortMixin, RBLNQwen2VLM
         inputs_sorted: bool = False,
         **kwargs,
     ) -> RBLNDecoderOnlyOutput:
-        self._check_batch_inputs_sorted(input_ids if input_ids is not None else inputs_embeds, inputs_sorted)
+        self._require_sorted_batch_inputs(input_ids if input_ids is not None else inputs_embeds, inputs_sorted)
         output_hidden_states = _validate_output_hidden_states(output_hidden_states, self.rbln_config)
         # Prefill
         if cache_position is None:
